preferred kernel: move to local.conf
x86-base.inc weak-sets the preferred kernel to linux-yocto, which is
parsed before the distro conf, and that overrides a weak-set in the
distro conf.
But the distro conf is parsed after local.conf so we can't use
a hard-set there or it can't be changed from local.conf without needing
an override.

Therefore just move it to local.conf all the time