initramfs: add a usable reboot/poweroff program
Busybox just signals init which does nothing in an initramfs shell, so
we need something functional to let reboot commands actually work.

There's a bit of logic to unmount and sync filesystems before issuing
the reboot syscall.

I did this in C++ mostly just to play around with the STL list/vector
when listing filesystems to unmount. Most of the APIs it calls are plain C.
I also got to learn update-alternatives Yocto infrastructure here.

This whole thing could probably be replaced with a shell script that
just writes to /proc/sysrq-trigger, but that's boring.