kernel: Finally, a working Arch-based config!
It turns out Arch really does configure *everything* as a module,
including USB and PS/2 keyboard support, loop devices, etc.

Drop the split between arch-to-yocto.cfg and towelstick.cfg, it wasn't
clear what should go where.
Drop the scc file and just use defconfig-arch.cfg as the kernel
defconfig (note the word 'defconfig' must be in the name for Yocto to be
happy using it)