remove busybox from the image
It was giving me some do_rootfs errors with the postinst script
(actually I think it was xfs not using update-alternatives properly),
but towelstick uses real coreutils and such so there's no need for
busybox anyway.

Remove the hard-coded busybox RDEPENDS in packagegroup-boot.

Force busybox in the initramfs instead of VIRTUAL-RUNTIME_base-utils
which is coreutils now.