tstick-image: enable qemuboot support
qemu is capable of booting from a raw disk image (in BIOS mode), and
runqemu works happily when there's no kernel in qemuboot.conf, but
do_write_qemuboot_conf always sets a kernel in the qemuboot.conf file.
Therefore I had to hack it a bit to remove all the kernel parameters
from the conf file.

Also the virtio-scsi driver is needed in the initramfs so that tstick
can find the root drive when running under qemu.