conf/machine: new tstickx86-64 machine which builds with -mtune=generic
The genericx86-64 machine uses tune core2-64, which sets
`-march=core2 -mtune=core2` in the target compile flags. Using a core2
architecture is good for compatibility, but it seems wrong to tune for
that CPU when towelstick will almost always be running on some other
processor generation.

Create a new core2-64-generic and tstickx86-64 machine which is
identical to core2-64/genericx896-64 except uses -mtune=generic instead
of -mtune=core2.

I don't know if this makes any practical difference, but it feels right
to do.